+++ lib/blkid/read.c
@@ -452,8 +452,8 @@ static void debug_dump_dev(blkid_dev dev)
 	}
 
 	printf("  dev: name = %s\n", dev->bid_name);
-	printf("  dev: DEVNO=\"0x%0llx\"\n", dev->bid_devno);
-	printf("  dev: TIME=\"%ld\"\n", dev->bid_time);
+	printf("  dev: DEVNO=\"0x%0llx\"\n", (long long)dev->bid_devno);
+	printf("  dev: TIME=\"%lld\"\n", (long long)dev->bid_time);
 	printf("  dev: PRI=\"%d\"\n", dev->bid_pri);
 	printf("  dev: flags = 0x%08X\n", dev->bid_flags);
